Refactor conference management by creating a new blueprint for conference routes and updating template links to use the new blueprint structure.
