+++ Frontend/templates/_results.html
@@ -31,7 +31,7 @@
                   {% endif %}
 
                   {% if session_user_name and session_user_name.userinfo %}
-                    <a href="{{ url_for('edit_conference', conf_id=article.id) }}"
+                    <a href="{{ url_for('conferences.edit_conference', conf_id=article.id) }}"
                       class="btn btn-warning btn-sm ms-2">
                       Edit
                     </a>
